:truck: move icons to Icon
