+++ my-code/src/components/Icon/index.js
@@ -1,0 +1,5 @@
+import Logo from './Logo'
+import Heart from './Heart'
+import Movie from './Movie'
+
+export { Logo, Heart, Movie }
